+++ lib/ui/components/title_bar_widget.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:photos/theme/ente_theme.dart';
-import 'package:photos/ui/components/icon_button_widget.dart';
+import 'package:photos/ui/components/buttons/icon_button_widget.dart';
 
 class TitleBarWidget extends StatelessWidget {
   final IconButtonWidget? leading;
